Better handling of no-tar param and file setup in the bowtie2 module

+++ tool/bowtie_aligner.py
@@ -110,8 +110,10 @@ class bowtie2AlignerTool(Tool):
             except IOError:
                 return False
 
-            gfl = genome_file_loc.split("/")
-            genome_fa_ln = genome_idx.replace('.tar.gz', '/') + gfl[-1]
+        gfl = genome_file_loc.split("/")
+        genome_fa_ln = genome_idx.replace('.tar.gz', '/') + gfl[-1]
+
+        if os.path.isfile(genome_fa_ln) is False:
             shutil.copy(genome_file_loc, genome_fa_ln)
 
         if (
@@ -186,8 +188,10 @@ class bowtie2AlignerTool(Tool):
             except IOError:
                 return False
 
-            gfl = genome_file_loc.split("/")
-            genome_fa_ln = genome_idx.replace('.tar.gz', '/') + gfl[-1]
+        gfl = genome_file_loc.split("/")
+        genome_fa_ln = genome_idx.replace('.tar.gz', '/') + gfl[-1]
+
+        if os.path.isfile(genome_fa_ln) is False:
             shutil.copy(genome_file_loc, genome_fa_ln)
 
         out_bam = read_file_loc1 + '.out.bam'
